requirements.txt
Flask==2.3.3
# transformers==4.36.2
replicate==1.0.4
python-dotenv==1.0.0
gunicorn==21.2.0
requests